Fix: 프리셋 MQTT 발행 로직 수정 및 기타 디버깅
- /preset_coordinate pub 로직 디버깅
- chatbot 및 backend 기타 버그 수정(자세 판단 로직)

#S13P11B101-234

+++ backend/siseon-backend/src/main/java/siseon/backend/dto/ProfileCreateRequest.java
@@ -19,7 +19,7 @@ public class ProfileCreateRequest {
     private LocalDate birthDate;
 
     @NotNull
-    private Float height;
+    private Integer height;
 
     @NotNull
     private Float leftVision;
